% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHUC.R
\name{getHUC}
\alias{getHUC}
\title{Obtain GIS and HUC inforamtion for new Stations}
\usage{
getHUC(sites, year)
}
\arguments{
\item{sites}{A list of Station Codes, stream names and coordinates.}

\item{year}{Label the year these site are to sampled, just used for naming export file.}
}
\value{
a csv file of StationCodes and coordinates is produced with GIS information necessary for creating new sites in SWAMP databse
}
\description{
Obtain GIS and HUC inforamtion for new Stations
}
\examples{
getHUC(sites, 2022)
required column names for sites are: StationCode, name, lat, long
}
